Move get_panel_name to own file
