+++ firmware/src/CMakeLists.txt
@@ -134,6 +134,7 @@ add_executable(
   ${FWDIR}/src/fs/helpers.cc
   ${FWDIR}/src/fw_update/updater_proxy.cc
   ${FWDIR}/src/gui/elements/element_name.cc
+  ${FWDIR}/src/gui/elements/panel_name.cc
   ${FWDIR}/src/gui/slsexport/ui_local.cc
   ${FWDIR}/src/gui/fonts/fonts.cc
   ${FWDIR}/src/gui/fonts/ttf.cc
