+++ firmware/src/gui/elements/element_name.cc
@@ -51,7 +51,7 @@ static std::string get_mapped_color(Element const &element, uint16_t panel_id) {
 }
 
 void append_panel_name(std::string &opts, Element const &el, uint16_t mapped_panel_id) {
-	const auto name = std::visit([=](auto &e) { return get_panel_name<PanelDef>(e, mapped_panel_id); }, el);
+	const auto name = std::visit([=](auto &e) { return get_panel_name(e, mapped_panel_id); }, el);
 	if (name.size() < 0)
 		return;
 	const auto color = get_mapped_color(el, mapped_panel_id);
@@ -74,7 +74,7 @@ void append_connected_jack_name(std::string &opts, GuiElement const &drawn, Patc
 		if (auto *cable = patch.find_internal_cable_with_injack(in_jack)) {
 			if (auto out_map = patch.find_mapped_outjack(cable->out)) {
 				const auto color = get_mapped_color(JackOutput{}, out_map->panel_jack_id);
-				const auto p_name = get_panel_name<PanelDef>(JackOutput{}, out_map->panel_jack_id);
+				const auto p_name = get_panel_name(JackOutput{}, out_map->panel_jack_id);
 				opts += " " + Gui::color_text(p_name, color);
 			}
