+++ simulator/CMakeLists.txt
@@ -135,6 +135,7 @@ else()
 		stubs/static_buffers.cc
 
 		${FWDIR}/src/gui/elements/element_name.cc
+		${FWDIR}/src/gui/elements/panel_name.cc
 		${FWDIR}/src/gui/slsexport/ui_local.cc
 		${FWDIR}/src/fs/helpers.cc
 		# ${FWDIR}/src/gui/slsexport/prefs_menu.cc
